--- app/Services/PatientService.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Services;
-
-use App\Models\Patient;
-
-class PatientService
-{
-    public function findPatientById($patientId)
-    {
-        return Patient::find($patientId);
-    }
-}
